% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{forest_plot_grid}
\alias{forest_plot_grid}
\title{Create a grid of forest plots}
\usage{
forest_plot_grid(
  ...,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  plot_grid_args = list()
)
}
\arguments{
\item{...}{Pass individual plots returned by forest_plot, or lists of such plots (bare lists will be spliced).}

\item{nrow, ncol}{Specify the grid (one is sufficient, uses auto layout if both are null)}

\item{byrow}{If the plots are given in by-row, or by-column (byrow=FALSE) order}

\item{plot_grid_args}{Additional arguments to the \code{\link[cowplot]{plot_grid}} function which is used to create the grid.}
}
\value{
Return value of \code{\link[cowplot]{plot_grid}}
}
\description{
Makes use of the stored layout information in a \code{\link{forest_plot}} plot to
create grids of plots.
}
