% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survminertools.R
\name{ggsurvplot_to_gtable}
\alias{ggsurvplot_to_gtable}
\title{Build a gtable representation from a ggsurvplot object}
\usage{
ggsurvplot_to_gtable(
  ggsurv_obj,
  surv.plot.height = NULL,
  risk.table.height = NULL,
  ncensor.plot.height = NULL
)
}
\arguments{
\item{ggsurv_obj}{The ggsurvplot object}

\item{surv.plot.height, risk.table.height, ncensor.plot.height}{Layout parameters, see \code{\link[survminer]{arrange_ggsurvplots}}}
}
\value{
A gtable object
}
\description{
Build a gtable representation from a ggsurvplot object
}
