% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenWoodError.R
\name{greenWoodError}
\alias{greenWoodError}
\title{Greenwood Estimate for Standard Error}
\usage{
greenWoodError(dataToUse, timeToMeasure)
}
\arguments{
\item{dataToUse}{The dataset used to create function}

\item{timeToMeasure}{Time at which we want to calculate survival probability}
}
\description{
This function calculates the Standard Error of our survival function
}
\examples{
greenWoodError()
}
