% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survminertools.R
\name{grid_layout}
\alias{grid_layout}
\title{Grid layouting}
\usage{
grid_layout(n, rows = NULL, cols = NULL)
}
\arguments{
\item{n}{Number of items in grid}

\item{rows, cols}{Pass one of rows or cols, or none, in which case auto layout is used.}
}
\value{
A numeric vector of length 2: rows, cols
}
\description{
Creates a grid layout nrow x ncol
for n items.
}
\examples{
grid_layout(24, cols=4)
grid_layout(24)
grid_layout(24, rows=2)
}
