% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{identity_order}
\alias{identity_order}
\title{Ordering function: identity order}
\usage{
identity_order(x, ...)
}
\arguments{
\item{x}{a vector}

\item{...}{Effectively ignored}
}
\value{
An integer vector
}
\description{
This can be used in a place where a function with a signature like \code{\link{order}} is required.
It simply retains the original order.
}
