% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survminertools.R
\name{in_one_kaplan_meier_plot}
\alias{in_one_kaplan_meier_plot}
\title{Display multiple survival curves within the same Kaplan Meier plot}
\usage{
in_one_kaplan_meier_plot(...)
}
\arguments{
\item{\dots}{Named SurvivalAnalysisResult objects as returned by \code{\link{analyse_survival}}.
Please note that the results need to come from the same data source and should
only differ by the "by" parameter. The first given result acts as source for common attributes.
Also takes one single named bare list.}
}
\value{
Internal object for use by \code{\link{kaplan_meier_plot}} or
\code{\link{kaplan_meier_grid}} only
}
\description{
Utility method to tell \code{\link{kaplan_meier_plot}} to combine the given
survival results within the same plot
}
