% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmSurvEst.R
\name{kmSurvEst}
\alias{kmSurvEst}
\title{Kaplan Meier Survival Estimate}
\usage{
kmSurvEst(dataToUse, timeToMeasure)
}
\arguments{
\item{dataToUse}{The dataset used to create function}

\item{timeToMeasure}{Time at which we want to calculate survival probability}
}
\description{
This function calculates the KM-Estimate of Survival Function
}
\examples{
kmSurvEst()
}
