% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmTruncConditional.R
\name{kmTruncConditional}
\alias{kmTruncConditional}
\title{Kaplan Meier Conditional Survival Estimate FOR LEFT TRUNCATED DATA}
\usage{
kmTruncConditional(dataToUse, timeToMeasure, conditionalTerm)
}
\arguments{
\item{dataToUse}{The dataset used to create function}

\item{timeToMeasure}{Time at which we want to calculate survival probability}

\item{conditionalTerm}{Time used as conditional term}
}
\description{
This function calculates the KM-Estimate of Conditional Survival Function with Truncation Considered
}
\examples{
kmTruncConditional()
}
