% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanSurvTime.R
\name{meanSurvTime}
\alias{meanSurvTime}
\title{Estimate for Mean Survival Time}
\usage{
meanSurvTime(dataToUse, maxTime)
}
\arguments{
\item{dataToUse}{The dataset used to create function}

\item{maxTime}{The upper bound of the range of time we will estimate mean time on.}
}
\description{
This function calculates the mean survival time given a upper bound time limit.
}
\examples{
meanSurvTime()
}
