% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naHazard.R
\name{naHazard}
\alias{naHazard}
\title{Nelson Aalen's Estimate for Hazard Function}
\usage{
naHazard(dataToUse, timeToMeasure)
}
\arguments{
\item{dataToUse}{The dataset used to create function}

\item{timeToMeasure}{Time at which we want to calculate survival probability}
}
\description{
This function calculates the NA Estimate of Hazard Function
}
\examples{
naHazard()
}
