% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naStdErrorHazard.R
\name{naStdErrorHazard}
\alias{naStdErrorHazard}
\title{Nelson Aalen Estimate for Standard Error}
\usage{
naStdErrorHazard(dataToUse, timeToMeasure)
}
\arguments{
\item{dataToUse}{The dataset used to create function}

\item{timeToMeasure}{Time at which we want to calculate survival probability}
}
\description{
This function calculates the Standard Error of our hazard function
}
\examples{
naStdErrorHazard()
}
