% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.R
\name{print.SurvivalAnalysisMultivariateResult}
\alias{print.SurvivalAnalysisMultivariateResult}
\title{Print the essentials of a SurvivalAnalySurvivalAnalysisMultivariateResult}
\usage{
\method{print}{SurvivalAnalysisMultivariateResult}(x, ..., p_precision = 3, hr_precision = 2, p_less_than_cutoff = 0.001)
}
\arguments{
\item{x}{The result generated by \code{\link{analyse_multivariate}}}

\item{...}{Further arguments passed from other methods.}

\item{p_precision, hr_precision}{Precision with which to print floating point values}

\item{p_less_than_cutoff}{Cut-off for small p values. Values smaller than this will be displayed like "<..."}
}
\value{
The formatted string, invisibly.
}
\description{
Print the essentials of a SurvivalAnalySurvivalAnalysisMultivariateResult
}
