% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{survival_essentials}
\alias{survival_essentials}
\title{Convenience formatting and printing of result}
\usage{
survival_essentials(
  result,
  label = NULL,
  p_precision = 3,
  hr_precision = 2,
  time_precision = 1,
  include_end_separator = TRUE,
  timespan_unit = "days",
  print = TRUE
)
}
\arguments{
\item{result}{The result generated by \code{\link{analyse_survival}}}

\item{label}{Optional label to include}

\item{p_precision, hr_precision, time_precision}{Precision with which to print floating point values}

\item{include_end_separator}{Append "\\n---\\n"? Comes handy if printing multiple results following each other}

\item{timespan_unit}{Unit for time spans: "days", "months" or "years".}

\item{print}{Print string to console}
}
\value{
The formatted string, invisibly. Ready for output with cat or saving to a file.
}
\description{
Takes the given result, formats and prints it
}
